refactor: mail preview API 파일 이동 #211
