+++ csalgo-server/src/test/java/kr/co/csalgo/server/mail/MailControllerTest.java
@@ -1,0 +1,23 @@
+package kr.co.csalgo.server.mail;
+
+import static org.springframework.test.web.servlet.request.MockMvcRequestBuilders.*;
+import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
+
+import org.junit.jupiter.api.Test;
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.MediaType;
+import org.springframework.test.web.servlet.MockMvc;
+
+class MailControllerTest {
+
+	@Autowired
+	private MockMvc mockMvc;
+
+	@Test
+	void testMailPreview() throws Exception {
+		mockMvc.perform(get("/mail-preview").accept(MediaType.TEXT_HTML))
+			.andExpect(status().isOk())
+			.andExpect(content().contentTypeCompatibleWith(MediaType.TEXT_HTML))
+			.andExpect(content().string(org.hamcrest.Matchers.containsString("시간 복잡도와 공간 복잡도를 설명하시오")));
+	}
+}
